Fix using hash expressions inside call expressions Fixes #326
