+++ expressions/arg.js
@@ -1,0 +1,13 @@
+// ### Arg
+// `new Arg(Expression [,modifierOptions] )`
+// Used to identify an expression that should return a value.
+var Arg = function(expression, modifiers){
+	this.expr = expression;
+	this.modifiers = modifiers || {};
+	this.isCompute = false;
+};
+Arg.prototype.value = function(){
+	return this.expr.value.apply(this.expr, arguments);
+};
+
+module.exports = Arg;
