+++ expressions/literal.js
@@ -1,0 +1,10 @@
+// ### Literal
+// For inline static values like `{{"Hello World"}}`
+var Literal = function(value){
+	this._value = value;
+};
+Literal.prototype.value = function(){
+	return this._value;
+};
+
+module.exports = Literal;
